Split out boundary methods into own Boundary class
No changes to the algorithm at this point, just moving from ScoutMap
